move utils higher
